Update taskme, don't quote builtins and &, &&.
If the command is an executable, you can have quotes around it. But
if it is a builtin like `copy` then quotes will break it. So let's
remove quotes from the command as long as it doesn't have spaces.

Also remove quotes from & and && because with quotes you can't actually
chain commands. However if you want to literally use & as an arg then
this may cause some problems.